Fix missing db links, change file extension, clean up documentation
Uses a relative path to always find the db correctly without having students enter the path themself.
Changed the file extension of the queries to .sqlite3-query so the SQLite3 Editor extension will detect them as queries for the database.
Rewrote some of the documentation strings, especially in student_queries to make it clearer and add missing information.